Moved test to the base class
