+++ collect_app/src/test/java/org/odk/collect/android/entities/DatabaseEntitiesRepositoryTest.kt
@@ -2,7 +2,6 @@ package org.odk.collect.android.entities
 
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.MatcherAssert.assertThat
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -36,20 +35,5 @@ class DatabaseEntitiesRepositoryTest : EntitiesRepositoryTest() {
 
         repository.save("things", savedEntity)
         assertThat(repository.getEntities("things")[0], sameEntityAs(savedEntity))
-    }
-
-    @Test
-    fun `#save ignores case-insensitive duplicate properties`() {
-        val repository = buildSubject()
-        val entity = Entity.New(
-            "1",
-            "One",
-            properties = listOf(Pair("prop", "value"), Pair("Prop", "value"))
-        )
-
-        repository.save("things", entity)
-        val savedEntities = repository.getEntities("things")
-        assertThat(savedEntities[0].properties.size, equalTo(1))
-        assertThat(savedEntities[0].properties[0].first, equalTo("prop"))
     }
 }
